+++ NEWS.md
@@ -5,6 +5,8 @@
      and `str_replace_all()` uses `REGEXP_REPLACE()` (#446).
      
   * `paste()` and `paste0()` use `||` (#458).
+  
+  * `as.numeric()` and `as.double()` cast to `FLOAT` (#408).
 
 * `sql_translate_env.Microsoft SQL Server()` now uses `sql_try_cast()` instead
   of `sql_cast()` for MSSQL version 11+ (2012+) (@DavidPatShuiFong, #380).